[PromotionBundle] Move logic from PromotionActionGroupValidator to PromotionActionValidator
