+++ src/Sylius/Bundle/CoreBundle/Resources/config/validation/PromotionAction.xml
@@ -13,10 +13,6 @@
 
 <constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
     <class name="Sylius\Component\Promotion\Model\PromotionAction">
-        <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\PromotionActionGroup">
-            <option name="groups">sylius_api</option>
-        </constraint>
-
         <property name="configuration">
             <constraint name="Collection">
                 <option name="groups">
@@ -49,6 +45,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection">
@@ -66,8 +63,10 @@
                                 </constraint>
                             </value>
                         </option>
+                        <option name="allowExtraFields">true</option>
                     </constraint>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection">
@@ -163,8 +162,10 @@
                                 </constraint>
                             </value>
                         </option>
+                        <option name="allowExtraFields">true</option>
                     </constraint>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection">
@@ -266,8 +267,10 @@
                                 </constraint>
                             </value>
                         </option>
+                        <option name="allowExtraFields">true</option>
                     </constraint>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
         </property>
     </class>
