+++ src/Sylius/Bundle/CoreBundle/Resources/config/validation/PromotionRule.xml
@@ -13,10 +13,6 @@
 
 <constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
     <class name="Sylius\Component\Promotion\Model\PromotionRule">
-        <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\PromotionRuleGroup">
-            <option name="groups">sylius_api</option>
-        </constraint>
-
         <property name="configuration">
             <constraint name="Collection">
                 <option name="groups">cart_quantity</option>
@@ -31,6 +27,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Collection">
@@ -49,6 +46,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Collection">
@@ -64,6 +62,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Collection">
@@ -82,6 +81,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Collection">
@@ -101,6 +101,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection">
@@ -127,8 +128,10 @@
                                 </constraint>
                             </value>
                         </option>
+                        <option name="allowExtraFields">true</option>
                     </constraint>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Collection">
@@ -147,6 +150,7 @@
                         </constraint>
                     </value>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
 
             <constraint name="Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection">
@@ -164,8 +168,10 @@
                                 </constraint>
                             </value>
                         </option>
+                        <option name="allowExtraFields">true</option>
                     </constraint>
                 </option>
+                <option name="allowExtraFields">true</option>
             </constraint>
         </property>
     </class>
